+++ Sources/AdaptyFlowKit/AdaptyFlowKit.swift
@@ -16,5 +16,5 @@
 /// ```
 public enum AdaptyFlowKit {
     /// Current SDK version
-    public static let version = "1.0.4"
+    public static let version = "1.0.5"
 }